setup: Have setuptools generate the run script
